# gateway/cs-api/openapi.yaml
openapi: 3.0.3
info:
  title: semconnect — OGC API Connected Systems (CS API) v0.1
  version: 0.1.0
  description: |-
    OAS3 contract for the cs-api gateway shipped by
    [semconnect](https://github.com/C360Studio/semconnect).

    cs-api exposes a subset of [OGC API Connected Systems](https://docs.ogc.org/DRAFTS/23-001r0.html)
    over the [semstreams](https://github.com/C360Studio/semstreams)
    framework primitives. v0.1 ships the endpoints listed below as
    implemented HTTP routes. The contract is intentionally narrower than
    the full upstream OGC OAS where v0.1 does not yet claim a resource or
    conformance class.

    **Honest deferral signals.** Several endpoints return data that is
    intentionally lossy or shape-truncated relative to the upstream
    spec. cs-api signals these via response headers:

    - `X-CS-Reconstructed-Lossy: true` on `GET /systems/{id}` — the
      SensorML / JSON-LD reconstruction from triples drops
      inputs/outputs/parameters/keywords/connections/identifier
      metadata that the framework's triple emitter does not preserve.
      Clients drilling into nested System children get skeleton refs.
    - `X-CS-Items-Count: <N>` on `GET /datastreams/{id}/observations`
      with `Accept: application/om+json` — count of items in the bare
      OMS array response (the JSON wrapper already carries
      `numberReturned`).

    **Auth posture.** Anonymous-by-default at v0.1. Optional
...

...
    Adapted from the upstream OGC OAS3 source at
...
servers:
...
security:
  - {}  # anonymous-by-default at v0.1 — middleware-body swap turns this on
paths:
...
  /systems/{id}:
    parameters:
      - $ref: "#/components/parameters/EntityID"
    get:
      operationId: getSystem
      summary: Fetch a system by ID.
      description: |-
        Returns the System resource in one of three encodings:

        - `application/json` — CS API §7.2 JSON subset (lossy, see header).
        - `application/sensorml+json` — SensorML reconstruction from
          triples (lossy).
        - `application/ld+json` — JSON-LD via the framework's RDF
          serializer.

        All three honestly signal lossy reconstruction via
        `X-CS-Reconstructed-Lossy: true`.
      tags: [Systems]
      parameters:
        - $ref: "#/components/parameters/F"
      responses:
        "200":
          description: System resource.
          headers:
            X-CS-Reconstructed-Lossy:
              $ref: "#/components/headers/CsReconstructedLossy"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/System"}
            application/sensorml+json:
              schema:
                type: object
                description: |-
                  Reconstructed SensorML PhysicalSystem / Sensor /
                  Procedure. Fields dropped by the triple-emitter
                  (inputs/outputs/parameters/keywords/connections) are
                  absent — see `X-CS-Reconstructed-Lossy` header.
            application/ld+json:
              schema:
                type: object
                description: JSON-LD with SOSA / SSN / SKOS context.
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
        "406": {$ref: "#/components/responses/NotAcceptable"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
    head:
      operationId: headSystem
      tags: [Systems]
      responses:
        "200": {description: OK (no body)}
        "404": {description: Not Found}
    put:
      operationId: replaceSystem
      summary: Replace a system (CS API §7.6 create-replace-delete).
      description: |-
        Replace-or-upsert semantics: existing triples for the entity
        are removed, then the request body's triples are written. If
        the entity does not exist, the body's triples are written
        fresh (CS API §7.6 standard PUT-as-upsert).

        The body's `properties.uid` must mint to the same entity ID
        as the path `{id}` — mismatch yields 400 *before* any
        destructive operation.

        Body media type is `application/json` or
        `application/geo+json` (GeoJSON Feature shape). SensorML is
        intentionally not accepted on PUT — the lossy triple
        round-trip would mismatch the read-back shape and surprise
        clients.
      tags: [Systems]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SystemFeature"}
          application/geo+json:
            schema: {$ref: "#/components/schemas/SystemFeature"}
      responses:
        "204":
          description: System replaced or upserted.
          headers:
            X-CS-Attempted-ID:
              schema: {type: string}
              description: Echoes the entity ID — present on every response so clients can correlate.
        "400": {$ref: "#/components/responses/BadRequest"}
        "409": {$ref: "#/components/responses/Conflict"}
        "413": {$ref: "#/components/responses/PayloadTooLarge"}
        "415": {$ref: "#/components/responses/UnsupportedMediaType"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
    patch:
      operationId: patchSystem
      summary: Partial-update a system (CS API conf/update).
      description: |-
        Partial-update semantics: each field present in the body
        replaces the corresponding triple on the entity; fields
        absent in the body are left alone. Body shape is the same
        SystemFeature schema PUT accepts (JSON or geo+json), but
        only `type: "Feature"` is enforced (when present) — all
        properties fields are optional, and the path `{id}` is
        authoritative.

        If the body specifies `properties.uid`, it MUST match the
        existing entity's preserved uid (Stage 18 sister-side
        triple). Mismatch yields 400 *before* any destructive
        operation.

        **No JSON Merge Patch null-as-delete** (RFC 7396) at v0.1:
        a `null` body field is treated as a no-op rather than a
        remove. The ETS does not exercise null-removes; the
        conservative stance avoids surprising existing clients.

        404 if the entity does not exist (no upsert; PUT is the
        upsert path).
      tags: [Systems]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SystemFeature"}
          application/geo+json:
            schema: {$ref: "#/components/schemas/SystemFeature"}
      responses:
        "204":
          description: System patched.
          headers:
            X-CS-Attempted-ID:
              schema: {type: string}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
        "413": {$ref: "#/components/responses/PayloadTooLarge"}
        "415": {$ref: "#/components/responses/UnsupportedMediaType"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
    delete:
      operationId: deleteSystem
      summary: Delete a system (CS API §7.6 create-replace-delete).
      description: |-
        Idempotent: a delete against a non-existent ID still returns
        204. Deletes the graph entity through the semstreams
        entity-level delete primitive.
      tags: [Systems]
      responses:
        "204": {description: System deleted (or never existed).}
        "400": {$ref: "#/components/responses/BadRequest"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
    options:
      operationId: optionsSystem
      summary: Advertise allowed methods on /systems/{id}.
      description: |-
        Returns 204 with an `Allow` header listing the methods the
        item endpoint supports. Stage 19 added PATCH (CS API
        `conf/update`).
      tags: [Systems]
      responses:
        "204":
          description: No Content.
          headers:
            Allow:
              schema: {type: string}
              description: '`GET, HEAD, PUT, PATCH, DELETE, OPTIONS`'
  /systems/{id}/subsystems:
    parameters:
      - $ref: "#/components/parameters/EntityID"
    get:
      operationId: getSystemSubsystems
      summary: List a system's subsystems.
      description: |-
        Stage 49 — CS API `/conf/subsystem` read-side collection. Returns
        child System entities whose stored SensorML relation points back to
        the parent with `sensorml.PredIsHostedBy`.
      tags: [Systems]
      parameters:
        - $ref: "#/components/parameters/F"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: SystemCollection of child systems.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SystemCollection"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
        "406": {$ref: "#/components/responses/NotAcceptable"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
    head:
      operationId: headSystemSubsystems
      tags: [Systems]
      responses:
        "200": {description: OK (no body)}
        "404": {description: Not Found}
  /systems/{id}/subsystems/{subsystemID}:
    parameters:
      - $ref: "#/components/parameters/EntityID"
      - $ref: "#/components/parameters/SubsystemID"
    get:
      operationId: getSystemSubsystem
      summary: Fetch a subsystem through its parent system.
      description: |-
        Subsystems are ordinary System resources. This endpoint first verifies
        that `{subsystemID}` is hosted by `{id}`, then returns the same
        encodings as `/systems/{subsystemID}`. The JSON representation includes
        `rel=canonical` and `rel=parent` links.
      tags: [Systems]
      parameters:
        - $ref: "#/components/parameters/F"
      responses:
        "200":
          description: Subsystem resource.
          headers:
            X-CS-Reconstructed-Lossy:
              $ref: "#/components/headers/CsReconstructedLossy"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/System"}
            application/sensorml+json:
              schema: {type: object}
            application/ld+json:
              schema: {type: object}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
        "406": {$ref: "#/components/responses/NotAcceptable"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
    head:
      operationId: headSystemSubsystem
      tags: [Systems]
      responses:
        "200": {description: OK (no body)}
        "404": {description: Not Found}

  # ============================================================
  # System History (OSH-bar vendor extension)
  # ============================================================
  /systems/{id}/history:
    x-cs-vendor-extension: true
    x-cs-upstream-source: api/upstream/part2/paths/systemHistory.yaml
    parameters:
      - $ref: "#/components/parameters/EntityID"
    get:
      operationId: getSystemHistory
      summary: List current system history.
      description: |-
        Stage 26 — OSH-bar System History read-side vendor extension.
        OGC 23-002 Annex A does not define a system-history
        conformance class in the pinned ETS, so this endpoint is
        documented but not claimed in `/conformance`.

        v0.1 exposes the current System description as the single
        available revision, linked as `current`.
      tags: [Systems]
      parameters:
        - $ref: "#/components/parameters/F"
      responses:
        "200":
          description: SystemCollection containing the current revision.
          headers:
            X-CS-History-Current-Only:
              schema: {type: string, enum: ["true"]}
              description: Indicates that only the current revision is available at v0.1.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SystemCollection"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
        "406": {$ref: "#/components/responses/NotAcceptable"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
    head:
      operationId: headSystemHistory
      tags: [Systems]
      responses:
        "200": {description: OK (no body)}
        "404": {description: Not Found}
    options:
      operationId: optionsSystemHistory
      tags: [Systems]
      responses:
        "204": {description: No Content.}
...
  /systems/{id}/datastreams:
    parameters:
      - $ref: "#/components/parameters/EntityID"
    get:
      operationId: getSystemDatastreams
      summary: List datastreams for a system.
      description: Stage 44 — filters Datastreams by the stored system reference.
      tags: [Datastreams]
      parameters:
        - $ref: "#/components/parameters/F"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: DatastreamCollection.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DatastreamCollection"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "406": {$ref: "#/components/responses/NotAcceptable"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
...
  /controlstreams/{id}/schema:
    parameters:
      - $ref: "#/components/parameters/EntityID"
    get:
      operationId: getControlStreamSchema
      tags: [ControlStreams]
      summary: Get a control stream command schema.
      responses:
        "200":
          description: Command schema.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CommandSchema"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
...